# BUILD : 1.7.15 (1135) # BUGS  : 586 584 582 # NOTES : several fixes
git-svn-id: svn://svn.anope.org/anope/trunk@1135 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@858 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/sockutil.c
@@ -725,7 +725,7 @@ char *ano_sockstrerror(int error)
 
 int ano_socksetnonb(ano_socket_t fd)
 {
-    int i = 1;
+    u_long i = 1;
     return (!ioctlsocket(fd, FIONBIO, &i) ? -1 : 1);
 }
 #endif
